requirements.txt
pytorch-lightning
torchmetrics
ftfy
matplotlib
tensorboard